FEATURE: Adds search for tickers
